requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
pydantic>=2.11.0
python-multipart>=0.0.20
opencv-contrib-python>=4.11.0
numpy>=1.26.0
psycopg2-binary>=2.9.10
alembic>=1.16.0
packaging>=25.0
pyyaml>=6.0.0
python-magic>=0.4.27
psutil>=7.0.0
Pillow>=11.2.0
pytest>=8.3.0
pytest-asyncio>=0.26.0
pytest-cov>=6.1.0
httpx>=0.28.0
watchdog>=6.0.0
jinja2>=3.1.0
requests>=2.32.0